Fixing bug where hold status was not settable.
The getStateStatuses call was returning a key/value array of code/label due to a default method parameter
that was adding labels. As we just care about the codes I changed the method call to not add labels.
Now the return from the method call is just an array of codes as expected/needed.